+++ paddlex/inference/common/result/mixin.py
@@ -581,160 +581,29 @@ class MarkdownMixin:
         self._markdown_writer = MarkdownWriter(*args, **kwargs)
         self._save_funcs.append(self.save_to_markdown)
 
-    def _to_markdown(self):
-        def _format_data(obj):
-
-            def format_title(content_value):
-                content_value = content_value.rstrip(".")
-                level = (
-                    content_value.count(
-                        ".",
-                    )
-                    + 1
-                    if "." in content_value
-                    else 1
-                )
-                return f"{'#' * level} {content_value}".replace("-\n", "").replace(
-                    "\n",
-                    " ",
-                )
-
-            def format_centered_text(key):
-                return (
-                    f'<div style="text-align: center;">{sub_block[key]}</div>'.replace(
-                        "-\n",
-                        "",
-                    ).replace("\n", " ")
-                    + "\n"
-                )
-
-            def format_image():
-                img_tags = []
-                if "img" in sub_block["image"]:
-                    img_tags.append(
-                        '<div style="text-align: center;"><img src="{}" alt="Image" /></div>'.format(
-                            sub_block["image"]["img"]
-                            .replace("-\n", "")
-                            .replace("\n", " "),
-                        ),
-                    )
-                if "image_text" in sub_block["image"]:
-                    img_tags.append(
-                        '<div style="text-align: center;">{}</div>'.format(
-                            sub_block["image"]["image_text"]
-                            .replace("-\n", "")
-                            .replace("\n", " "),
-                        ),
-                    )
-                return "\n".join(img_tags)
-
-            def format_chart():
-                img_tags = []
-                if "img" in sub_block["chart"]:
-                    img_tags.append(
-                        '<div style="text-align: center;"><img src="{}" alt="Image" /></div>'.format(
-                            sub_block["chart"]["img"]
-                            .replace("-\n", "")
-                            .replace("\n", " "),
-                        ),
-                    )
-                if "image_text" in sub_block["chart"]:
-                    img_tags.append(
-                        '<div style="text-align: center;">{}</div>'.format(
-                            sub_block["chart"]["image_text"]
-                            .replace("-\n", "")
-                            .replace("\n", " "),
-                        ),
-                    )
-                return "\n".join(img_tags)
-
-            def format_reference():
-                pattern = r"\[\d+\]"
-                res = re.sub(
-                    pattern,
-                    lambda match: "\n" + match.group(),
-                    sub_block["reference"],
-                )
-                return "\n" + res
-
-            def format_table():
-                return "\n" + sub_block["table"]
-
-            handlers = {
-                "paragraph_title": lambda: format_title(sub_block["paragraph_title"]),
-                "doc_title": lambda: f"# {sub_block['doc_title']}".replace(
-                    "-\n",
-                    "",
-                ).replace("\n", " "),
-                "table_title": lambda: format_centered_text("table_title"),
-                "figure_title": lambda: format_centered_text("figure_title"),
-                "chart_title": lambda: format_centered_text("chart_title"),
-                "text": lambda: sub_block["text"].strip("\n"),
-                # 'number': lambda: str(sub_block['number']),
-                "abstract": lambda: "\n" + sub_block["abstract"].strip("\n"),
-                "content": lambda: sub_block["content"]
-                .replace("-\n", "")
-                .replace("\n", " ")
-                .strip(),
-                "image": format_image,
-                "chart": format_chart,
-                "formula": lambda: f"$${sub_block['formula']}$$".replace(
-                    "-\n",
-                    "",
-                ).replace("\n", " "),
-                "table": format_table,
-                "reference": format_reference,
-                "algorithm": lambda: "\n"
-                + f"**Algorithm**: {sub_block['algorithm']}".replace("-\n", "").replace(
-                    "\n",
-                    " ",
-                ),
-                "seal": lambda: "\n"
-                + f"**Seal**: {sub_block['seal']}".replace("-\n", "").replace(
-                    "\n",
-                    " ",
-                ),
-            }
-            parsing_result = obj["layout_parsing_result"]
-            markdown_content = ""
-            for block in parsing_result:  # for each block show ordering results
-                sub_blocks = block["sub_blocks"]
-                last_label = None
-                seg_start_flag = None
-                seg_end_flag = None
-                for sub_block in sorted(
-                    sub_blocks,
-                    key=lambda x: x.get("sub_index", 999),
-                ):
-                    label = sub_block.get("label")
-                    seg_start_flag = sub_block.get("seg_start_flag")
-                    handler = handlers.get(label)
-                    if handler:
-                        if (
-                            label == last_label == "text"
-                            and seg_start_flag == seg_end_flag == False
-                        ):
-                            markdown_content += " " + handler()
-                        else:
-                            markdown_content += "\n\n" + handler()
-                        last_label = label
-                        seg_end_flag = sub_block.get("seg_end_flag")
-
-            return markdown_content
-
-        return _format_data(self)
+    def _to_markdown(self, save_path=None):
+        """
+        Convert the result to markdown format.
+        Args:
+            save_path: str
+        Returns:
+            Dict
+        """
+        raise NotImplementedError
 
     @property
-    def markdown(self):
-        return self._to_markdown()
+    def markdown(self, save_path=None):
+        return self._to_markdown(save_path)
 
     def save_to_markdown(self, save_path, *args, **kwargs):
-        if not str(save_path).endswith(".md"):
-            save_path = Path(save_path) / f"{Path(self['input_path']).stem}.md"
+        save_path = Path(save_path)
+        if not save_path.suffix.lower() == ".md":
+            save_path = save_path / f"layout_parsing_result.md"
+
         self._save_list_data(
             self._markdown_writer.write,
             save_path,
-            self.markdown,
+            self.markdown(save_path),
             *args,
             **kwargs,
         )
